chore: rename world-player.entity to world-player.component
 - the entity stuff is confusing. It's just another angular component.
