+++ src/app/routes/world/world-player.component.ts
@@ -29,29 +29,7 @@ import { MapFeatureComponent } from './map-feature.component';
 @Component({
   selector: 'world-player',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  template: `
-    <player-render-behavior #render></player-render-behavior>
-    <collision-behavior #collision></collision-behavior>
-    <player-map-path-behavior [tileMap]="map" #path></player-map-path-behavior>
-    <player-behavior
-      (onCompleteMove)="encounter.completeMove($event)"
-      [map]="map"
-      #player
-    ></player-behavior>
-    <combat-encounter-behavior
-      [scene]="scene"
-      #encounter
-      [tileMap]="map"
-      [player]="self"
-    ></combat-encounter-behavior>
-    <player-camera-behavior #camera></player-camera-behavior>
-    <player-look-behavior
-      (onLook)="onFeatureLook($event)"
-      (onLookAway)="onFeatureLookAway($event)"
-      #trigger
-    ></player-look-behavior>
-    <ng-content></ng-content>
-  `,
+  templateUrl: 'world-player.component.html',
 })
 export class WorldPlayerComponent
   extends GameEntityObject
